Move handling of query params into onboard store into login step
